moved string-avatar to components
