+++ client/src/components/string-avatar/component.tsx
