+++ client/src/components/string-avatar/index.ts
