+++ client/src/components/string-avatar/type.d.ts
